+++ R/OR.R
@@ -1,11 +1,15 @@
-#' Alternative return for statements
+#' Nullish coalescing operator
 #'
-#' Return alternative if the value of expression is empty or NA or NULL
+#' Alternative return for empty, null or na statements.Return alternative if the value of expression is empty or NA or NULL
 #'
 #' @rdname orsign
 #' @param test an object to return
 #' @param alternative alternative object to return
 #' @return value of test if not null or empty, else return value of alternative
+#' @note
+#' Equivalent to Nullish coalescing operator ?? in javascript or PHP like $Var = $operand1 ?? $operand2;
+#'
+#' @examples
 #'
 #' test1 <- c(4,NA,5,2,0,21)
 #'
@@ -39,3 +43,37 @@ or <- function(test,alternative){
 #' @rdname orsign
 #' @export
 `%or%` <- or
+
+#' Error coalescing operator
+#'
+#' Alternative return for error statements.Return alternative if the value of expression is erroneous
+#'
+#' @param test an object to return
+#' @param alternative alternative object to return
+#' @return value of test if error, else return value of alternative
+#' @rdname errorout
+#'
+#' @examples
+#' # The following statement would produce
+#' # error because 'stat1' does not exist
+#'
+#' # stat1 + 1
+#'
+#' # To prevent the statement from
+#' # stopping the process, when can have alternative out
+#' alt = 0
+#' error.out(stats1 + 1, alt)
+#'
+#' @export
+error.out <- function(test,alternative = ""){
+  tryCatch({
+    test
+  },error = function(e){
+    alternative
+  })
+}
+
+#' @rdname errorout
+#' @export
+`%eo%` <- error.out
+